+++ common/src/test/java/com/yanny/ali/platform/TestPlatformHelper.java
@@ -3,6 +3,8 @@ package com.yanny.ali.platform;
 import com.yanny.ali.api.IPlugin;
 import com.yanny.ali.platform.services.IPlatformHelper;
 import com.yanny.ali.plugin.Plugin;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.item.SpawnEggItem;
 import net.minecraft.world.level.storage.loot.LootPool;
 import net.minecraft.world.level.storage.loot.LootTable;
 
@@ -22,6 +24,11 @@ public class TestPlatformHelper implements IPlatformHelper {
 
     @Override
     public Path getConfiguration() {
+        return null;
+    }
+
+    @Override
+    public SpawnEggItem getSpawnEggItem(EntityType<?> entityType) {
         return null;
     }
 }